refactor authentication into its own package
