+++ cpm-hub/authentication/authenticator.h
