+++ cpm-hub/authentication/basic_authenticator.cpp
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 #include <boost/algorithm/string.hpp>
-#include <users/basic_authenticator.h>
+#include <authentication/basic_authenticator.h>
 
 using namespace std;
