+++ cpm-hub/management/deploy_service.h
@@ -21,7 +21,7 @@
 #include <vector>
 
 #include <infrastructure/filesystem.h>
-#include <users/authenticator.h>
+#include <authentication/authenticator.h>
 
 class DeployService {
 public:
