+++ tests/unit/plugins/test_plugin_index.cpp
