+++ tests/unit/plugins/test_plugins_api.cpp
