+++ tests/unit/plugins/test_plugins_repository_in_memory.cpp
