+++ tests/unit/plugins/test_plugins_service.cpp
